+++ app/db/postgres/PostgresStreams.scala
@@ -268,4 +268,32 @@ class PostgresStreams @Inject() (db: Database, sensors: Sensors) extends Streams
       st.close()
     }
   }
+
+  def getCount(sensor_id: Option[Int]): Int = {
+    var output = 0
+    db.withConnection { conn =>
+      var query = "SELECT COUNT(*) FROM streams"
+      sensor_id match {
+        case Some(id) => {
+          query += " WHERE sensor_id = ?"
+        }
+        case None =>
+      }
+      val st = conn.prepareStatement(query)
+      sensor_id match {
+        case Some(id) => {
+          st.setInt(1, id)
+        }
+        case None =>
+      }
+      val rs = st.executeQuery()
+      while (rs.next()) {
+        output = rs.getInt(1)
+      }
+      rs.close()
+      st.close()
+    }
+    output
+  }
+
 }