+++ CHANGELOG.md
@@ -29,6 +29,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-1134](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1134)
 - Added endpoint for analysis trends
   [GEOD-1062](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1062)
+- Implemented counts endpoint for sensors, streams, datapoints and bins
+  [GEOD-1154](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1154)
   
 ### Changed
 - Changed header and email title